Handle local labels correctly
All labels are currently treated as unique global symbols and are
suggested as completions anywhere in the project.

Labels with a dot prefix are contextual to their preceding regular
label. The same local label name can be defined multiple times in
different contexts. It only makes sense to offer local labels as a
completion when in the relevant scope.

The simplest solution for all of this is to prefix local labels when
defining and looking up symbols, effectively name-spacing them to their
parent label.

e.g.
```
Main:
.example
```
would be stored internally as `Main.example`