Integrate stand-alone debug adapter #219
The debugger logic is now in a separate
[uae-dap](https://github.com/grahambates/uae-dap/) package, allowing
it to be used with other editors that support Microsoft's [Debug Adapter
Protocol](https://microsoft.github.io/debug-adapter-protocol/).

In this change I removed all of the logic now handled by the adapter
package (along with related tests) and add the package as a dependency.
In order to retain VS Code specific functionality, I extend the session
classes and override a couple of methods to introduce the required hooks.

As an upshot of these changes, the extension no longer requires the
`cstool` binary, as this is included as wasm in the package.